+++ RELEASENOTES.md
@@ -4,9 +4,14 @@
 
 ## Release 0.1.0
 
+**BREAKING CHANGES**
+- Rename ContextBar to GlobalNavigationBar
+
 **MAJOR CHANGES**
 - Add the Global Header component
 - Add a Search component
+- Updates BreadCrumb to the latest markup
+- Updates PageHeader to the latest markup
 
 **OTHER**
 - Begin counting versions from 0.1.0